feat(options): 13+图表提供获取默认配置项的静态方法 (#2489)
* feat(heatmap): 热力图提供获取默认配置项的静态方法

- [x] testcases

* refactor(bullet): 获取子弹图的默认配置项

* refactor(histogram): 获取直方图的默认配置项

* refactor(rose): 获取玫瑰图的默认配置项

* refactor: 批量文件重命名

* refactor(sunburst): 获取旭日图的默认配置项

* refactor(scatter): 获取散点图的默认配置项

* refactor(wordcloud): 获取词云图的默认配置项

* refactor(waterfall): 获取瀑布图的默认配置项

* refactor(stock): 获取股票图的默认配置项

* refactor(radial-bar): 获取玉珏图的默认配置项

* refactor(funnel): 获取漏斗图的默认配置项

* refactor(chord): 获取弦图的默认配置项

* refactor(box): 获取箱形图的默认配置项

* refactor(bi-directional-bar): 获取对称条形图的默认配置项

* test: 增加单测
