+++ __tests__/unit/plots/heatmap/index-spec.ts
@@ -2,6 +2,7 @@ import { Heatmap } from '../../../../src';
 import { basicHeatmapData, semanticBasicHeatmapData } from '../../../data/basic-heatmap';
 import { createDiv } from '../../../utils/dom';
 import { DEFAULT_COLORS } from '../../../../src/constant';
+import { DEFAULT_OPTIONS } from '../../../../src/plots/heatmap/constant';
 
 describe('heatmap', () => {
   it('x*y with color', () => {
@@ -15,6 +16,9 @@ describe('heatmap', () => {
     });
 
     heatmap.render();
+
+    // @ts-ignore
+    expect(heatmap.getDefaultOptions()).toBe(Heatmap.getDefaultOptions());
 
     const geometry = heatmap.chart.geometries[0];
 
@@ -93,5 +97,9 @@ describe('heatmap', () => {
     expect(elements[minElementIndex].getModel().color.toUpperCase()).toBe(colors[0]);
 
     heatmap.destroy();
+  });
+
+  it('defaultOptions 保持从 constants 中获取', () => {
+    expect(Heatmap.getDefaultOptions()).toEqual(DEFAULT_OPTIONS);
   });
 });
