+++ src/plots/chord/index.ts
@@ -1,8 +1,8 @@
 import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
-import { Datum } from '../../types';
 import { ChordOptions } from './types';
 import { adaptor } from './adaptor';
+import { DEFAULT_OPTIONS } from './constant';
 
 export type { ChordOptions };
 
@@ -10,54 +10,18 @@ export type { ChordOptions };
  *  弦图 Chord
  */
 export class Chord extends Plot<ChordOptions> {
+  /**
+   * 获取 面积图 默认配置项
+   * 供外部使用
+   */
+  static getDefaultOptions(): Partial<ChordOptions> {
+    return DEFAULT_OPTIONS;
+  }
   /** 图表类型 */
   public type: string = 'chord';
 
   protected getDefaultOptions() {
-    return {
-      nodeStyle: {
-        opacity: 1,
-        fillOpacity: 1,
-        lineWidth: 1,
-      },
-      edgeStyle: {
-        opacity: 0.5,
-        lineWidth: 2,
-      },
-      label: {
-        fields: ['x', 'name'],
-        callback: (x: number[], name: string) => {
-          const centerX = (x[0] + x[1]) / 2;
-          const offsetX = centerX > 0.5 ? -4 : 4;
-          return {
-            labelEmit: true,
-            style: {
-              fill: '#8c8c8c',
-            },
-            offsetX,
-            content: name,
-          };
-        },
-      },
-      tooltip: {
-        fields: ['source', 'target', 'value'],
-        formatter: (datum: Datum) => {
-          const { source, target, value } = datum;
-          return {
-            name: `${source} -> ${target}`,
-            value,
-          };
-        },
-      },
-      interactions: [
-        {
-          type: 'element-active',
-        },
-      ],
-      weight: true,
-      nodePaddingRatio: 0.1,
-      nodeWidthRatio: 0.05,
-    };
+    return Chord.getDefaultOptions();
   }
 
   /**
