+++ src/plots/histogram/constant.ts
@@ -1,0 +1,17 @@
+import { Plot } from '../../core/plot';
+import { deepAssign } from '../../utils';
+
+/**
+ * 默认配置项
+ */
+export const DEFAULT_OPTIONS = deepAssign({}, Plot.getDefaultOptions(), {
+  // @ts-ignore
+  columnStyle: {
+    stroke: '#FFFFFF',
+  },
+  tooltip: {
+    shared: true,
+    showMarkers: false,
+  },
+  interactions: [{ type: 'active-region' }],
+});
