+++ src/plots/radial-bar/index.ts
@@ -1,8 +1,8 @@
 import { Plot } from '../../core/plot';
-import { deepAssign } from '../../utils';
 import { Adaptor } from '../../core/adaptor';
 import { RadialBarOptions } from './types';
 import { adaptor, meta } from './adaptor';
+import { DEFAULT_OPTIONS } from './constant';
 
 export type { RadialBarOptions };
 
@@ -10,6 +10,9 @@ export type { RadialBarOptions };
  * 玉珏图
  */
 export class RadialBar extends Plot<RadialBarOptions> {
+  static getDefaultOptions(): Partial<RadialBarOptions> {
+    return DEFAULT_OPTIONS;
+  }
   /** 图表类型 */
   public type: string = 'radial-bar';
 
@@ -28,19 +31,7 @@ export class RadialBar extends Plot<RadialBarOptions> {
    * 获取默认配置
    */
   protected getDefaultOptions(): Partial<RadialBarOptions> {
-    return deepAssign({}, super.getDefaultOptions(), {
-      interactions: [{ type: 'element-active' }],
-      legend: false,
-      tooltip: {
-        showMarkers: false,
-      },
-      xAxis: {
-        grid: null,
-        tickLine: null,
-        line: null,
-      },
-      maxAngle: 240,
-    });
+    return RadialBar.getDefaultOptions();
   }
 
   /**
