+++ src/plots/stock/constant.ts
@@ -1,3 +1,6 @@
+import { Plot } from '../../core/plot';
+import { deepAssign } from '../../utils';
+
 export const Y_FIELD = '$$stock-range$$';
 
 export const TREND_FIELD = 'trend';
@@ -21,3 +24,15 @@ export const DEFAULT_TOOLTIP_OPTIONS = {
     '<span style="background-color:{color};" class="g2-tooltip-marker"></span>' +
     '{name}{value}</li>',
 };
+
+/**
+ * 散点图 默认配置项
+ */
+export const DEFAULT_OPTIONS = deepAssign({}, Plot.getDefaultOptions(), {
+  // 设置默认图表 tooltips
+  tooltip: DEFAULT_TOOLTIP_OPTIONS,
+  interactions: [{ type: 'tooltip' }],
+  legend: {
+    position: 'top-left',
+  },
+});
