+++ src/plots/stock/index.ts
@@ -1,15 +1,21 @@
-import { deepAssign } from '../../utils';
 import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
 import { StockOptions } from './types';
 import { adaptor } from './adaptor';
 import { getStockData } from './utils';
-
-import { DEFAULT_TOOLTIP_OPTIONS } from './constant';
+import { DEFAULT_OPTIONS } from './constant';
 
 export type { StockOptions };
 
 export class Stock extends Plot<StockOptions> {
+  /**
+   * 获取 散点图 默认配置项
+   * 供外部使用
+   */
+  static getDefaultOptions(): Partial<StockOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public type: string = 'stock';
 
@@ -18,14 +24,7 @@ export class Stock extends Plot<StockOptions> {
    *  g2/g2plot默 认 配 置 -->  图 表 默 认 配 置  --> 开 发 者 自 定 义 配 置  --> 最 终 绘 图 配 置
    */
   protected getDefaultOptions(): Partial<StockOptions> {
-    return deepAssign({}, super.getDefaultOptions(), {
-      // 设置默认图表 tooltips
-      tooltip: DEFAULT_TOOLTIP_OPTIONS,
-      interactions: [{ type: 'tooltip' }],
-      legend: {
-        position: 'top-left',
-      },
-    });
+    return Stock.getDefaultOptions();
   }
 
   /**
