+++ src/plots/waterfall/index.ts
@@ -2,6 +2,7 @@ import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
 import { WaterfallOptions } from './types';
 import { adaptor } from './adaptor';
+import { DEFAULT_OPTIONS } from './constant';
 import { transformData } from './utils';
 
 export type { WaterfallOptions };
@@ -10,6 +11,14 @@ export type { WaterfallOptions };
  * 瀑布图
  */
 export class Waterfall extends Plot<WaterfallOptions> {
+  /**
+   * 获取 瀑布图 默认配置项
+   * 供外部使用
+   */
+  static getDefaultOptions(): Partial<WaterfallOptions> {
+    return DEFAULT_OPTIONS;
+  }
+
   /** 图表类型 */
   public readonly type: string = 'waterfall';
 
@@ -34,39 +43,6 @@ export class Waterfall extends Plot<WaterfallOptions> {
    * 获取 瀑布图 的默认配置
    */
   protected getDefaultOptions(): Partial<WaterfallOptions> {
-    return {
-      /** default: show label */
-      label: {},
-      /** default: show leaderLine */
-      leaderLine: {
-        style: {
-          lineWidth: 1,
-          stroke: '#8c8c8c',
-          lineDash: [4, 2],
-        },
-      },
-      /** default: show total */
-      total: {
-        label: '总计',
-        style: {
-          fill: 'rgba(0, 0, 0, 0.25)',
-        },
-      },
-      interactions: [{ type: 'element-active' }],
-      risingFill: '#f4664a',
-      fallingFill: '#30bf78',
-      waterfallStyle: {
-        fill: 'rgba(0, 0, 0, 0.25)',
-      },
-      yAxis: {
-        grid: {
-          line: {
-            style: {
-              lineDash: [4, 2],
-            },
-          },
-        },
-      },
-    };
+    return Waterfall.getDefaultOptions();
   }
 }
